requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
duckdb==0.9.2
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6